Move PeakFitFactory into correct subpackage refs #219
