+++ pyrs/core/pyrscore.py
@@ -2,16 +2,13 @@
 from pyrs.utilities import checkdatatypes
 from pyrs.core import instrument_geometry
 from pyrs.utilities import file_util
-from pyrs.core import peak_fit_factory
+from pyrs.peaks import PeakFitEngineFactory, SupportedPeakProfiles, SupportedBackgroundTypes
 from pyrs.utilities.rs_project_file import HidraConstants, HidraProjectFile, HidraProjectFileMode
 from pyrs.core import strain_stress_calculator
 from pyrs.core import reduction_manager
 from pyrs.core import polefigurecalculator
 import os
 import numpy
-
-# Define Constants
-SUPPORTED_PEAK_TYPES = ['PseudoVoigt', 'Gaussian', 'Voigt']  # 'Lorentzian': No a profile of HB2B
 
 
 class PyRsCore(object):
@@ -108,8 +105,7 @@ class PyRsCore(object):
         # get workspace
         workspace = self.reduction_service.get_hidra_workspace(fit_tag)
         # create a controller from factory
-        self._peak_fitting_dict[fit_tag] = peak_fit_factory.PeakFitEngineFactory.getInstance('Mantid')(workspace,
-                                                                                                       None)
+        self._peak_fitting_dict[fit_tag] = PeakFitEngineFactory.getInstance('Mantid')(workspace, None)
         # set wave length: TODO - #81+ - shall be a way to use calibrated or non-calibrated
         wave_length_dict = workspace.get_wavelength(calibrated=False, throw_if_not_set=False)
         if wave_length_dict is not None:
@@ -154,10 +150,8 @@ class PyRsCore(object):
 
         # Check Inputs
         checkdatatypes.check_dict('Peak fitting (information) parameters', peaks_fitting_setup)
-        checkdatatypes.check_string_variable('Peak type', peak_type,
-                                             peak_fit_factory.SupportedPeakProfiles)
-        checkdatatypes.check_string_variable('Background type', background_type,
-                                             peak_fit_factory.SupportedBackgroundTypes)
+        checkdatatypes.check_string_variable('Peak type', peak_type, SupportedPeakProfiles)
+        checkdatatypes.check_string_variable('Background type', background_type, SupportedBackgroundTypes)
 
         # Deal with sub runs
         if sub_run_list is None:
@@ -647,4 +641,4 @@ class PyRsCore(object):
         list of supported peaks' types for fitting
         :return:
         """
-        return SUPPORTED_PEAK_TYPES[:]
+        return SupportedPeakProfiles[:]
