+++ pyrs/peaks/__init__.py
@@ -1,3 +1,7 @@
+# flake8: noqa
 from __future__ import (absolute_import, division, print_function)  # python3 compatibility
 
-from .peak_collection import PeakCollection
+from .peak_collection import *
+from .peak_fit_factory import *
+
+__all__ = peak_collection.__all__ + peak_fit_factory.__all__
